separated appointments functions into a new module, fixed bug with diagnoses
